pyproject.toml
[tool.poetry]
name = "add-noggles"
version = "0.1.0"
description = ""
authors = ["akirawuc <stu970029@gmail.com>"]
readme = "README.md"
packages = [{include = "add_noggles"}]

[tool.poetry.dependencies]
python = "3.10.12"
opencv-python = "^4.7.0.72"
google-cloud-vision = "^3.4.2"
pillow = "^9.5.0"
flask = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
